Refactor condition builder content and rule builder in service store
